feat(core): Split config/errors/types/schemas and move identity-resolution + context-establishment-hooks into identity-substrate (no-changelog)
Fourth step of the token-exchange/identity-substrate split (IAM-1181).

- Splits TokenExchangeConfig into IdentitySubstrateConfig (trustedKeys,
  keyRefreshIntervalSeconds, jtiCleanupIntervalSeconds/BatchSize,
  inboundAudience, inboundSubjectClaim, excludeOwner) and a trimmed
  TokenExchangeConfig (enabled, embedEnabled, maxTokenTtl, rate limits).
  Env var names are unchanged - only which class declares them.
- Moves token-exchange.errors.ts to identity-substrate.errors.ts
  unchanged (IdentityResolutionService already throws these and can't
  depend on the consumer module).
- Splits token-exchange.types.ts: TokenExchangeFailureReason moves to
  identity-substrate.types.ts (nearly every value is substrate-produced);
  token-exchange.types.ts re-exports it so existing call sites keep
  working, and keeps its own consumer-only shapes (IssuedJwtPayload,
  TokenExchangeAuditEvent, ...).
- Splits token-exchange.schemas.ts: substrate-owned schemas
  (ExternalTokenClaims, TrustedKeySource, TrustedKeyData,
  ResolvedTrustedKey, ...) move to identity-substrate.schemas.ts;
  TokenExchangeRequestSchema stays.
- Moves context-establishment-hooks/{inbound-audience.service,
  inbound-claim-verification-hook} and identity-resolution.service.ts
  into identity-substrate.

Still no module-boundary change - token-exchange.module.ts's init()
still does every registration, just importing the moved substrate
pieces cross-module. Behaviorally a no-op.
