+++ packages/cli/src/modules/identity-substrate/context-establishment-hooks/__tests__/inbound-claim-verification-hook.test.ts
@@ -5,8 +5,8 @@ import type { Mocked } from 'vitest';
 
 import type { ExternalTokenVerifierProxy } from '@/services/external-token-verifier-proxy.service';
 
-import { InboundClaimVerificationHook } from '../inbound-claim-verification-hook';
 import type { InboundAudienceService } from '../inbound-audience.service';
+import { InboundClaimVerificationHook } from '../inbound-claim-verification-hook';
 
 function triggerItems(headers?: Record<string, unknown>): INodeExecutionData[] {
 	return [{ json: { headers } }];
