+++ packages/cli/src/modules/identity-substrate/identity-substrate.errors.ts
@@ -1,7 +1,7 @@
 import { AuthError } from '@/errors/response-errors/auth.error';
 import { BadRequestError } from '@/errors/response-errors/bad-request.error';
 
-import type { TokenExchangeFailureReason } from './token-exchange.types';
+import type { TokenExchangeFailureReason } from './identity-substrate.types';
 
 /**
  * AuthError subclass that carries a typed failure reason for Prometheus metrics.
