+++ packages/cli/src/modules/identity-substrate/services/__tests__/identity-resolution.service.test.ts
@@ -9,16 +9,16 @@ import {
 import { mock } from 'vitest-mock-extended';
 
 import type { EventService } from '@/events/event.service';
-import type { TrustedKeyService } from '@/modules/identity-substrate/services/trusted-key.service';
 import { IdentityBindingService } from '@/services/identity-binding.service';
 import type { RoleService } from '@/services/role.service';
 import type { UserService } from '@/services/user.service';
 
-import type { TokenExchangeConfig } from '../../token-exchange.config';
-import { TokenExchangeAuthError } from '../../token-exchange.errors';
-import type { ExternalTokenClaims } from '../../token-exchange.schemas';
-import { TokenExchangeFailureReason } from '../../token-exchange.types';
+import type { IdentitySubstrateConfig } from '../../identity-substrate.config';
+import { TokenExchangeAuthError } from '../../identity-substrate.errors';
+import type { ExternalTokenClaims } from '../../identity-substrate.schemas';
+import { TokenExchangeFailureReason } from '../../identity-substrate.types';
 import { IdentityResolutionService, qualifiedProviderId } from '../identity-resolution.service';
+import type { TrustedKeyService } from '../trusted-key.service';
 
 const logger = mock<Logger>({ scoped: vi.fn().mockReturnThis() });
 const userRepository = mock<UserRepository>();
@@ -27,7 +27,7 @@ const eventService = mock<EventService>();
 const userService = mock<UserService>();
 const trustedKeyService = mock<TrustedKeyService>();
 const roleService = mock<RoleService>();
-const config = mock<TokenExchangeConfig>();
+const config = mock<IdentitySubstrateConfig>();
 
 // The shared core is wired for real, against mocked repositories, so these
 // stay end-to-end regression tests of the resolution behaviour rather than
