+++ packages/cli/src/modules/identity-substrate/services/__tests__/trusted-key-sync.service.test.ts
@@ -6,11 +6,10 @@ import { jsonParse } from 'n8n-workflow';
 import { createHash } from 'node:crypto';
 import { mock } from 'vitest-mock-extended';
 
-import type { TokenExchangeConfig } from '@/modules/token-exchange/token-exchange.config';
-
 import { TrustedKeySourceEntity } from '../../database/entities/trusted-key-source.entity';
 import { TrustedKeyEntity } from '../../database/entities/trusted-key.entity';
 import type { TrustedKeySourceRepository } from '../../database/repositories/trusted-key-source.repository';
+import type { IdentitySubstrateConfig } from '../../identity-substrate.config';
 import type { JwksResolverService } from '../jwks-resolver';
 import { TrustedKeySyncService } from '../trusted-key-sync.service';
 
@@ -47,7 +46,7 @@ function createMocks({
 	inboundRequireVerifiedEmail?: boolean;
 	ssoInboundAudiences?: string;
 } = {}) {
-	const config = mock<TokenExchangeConfig>({
+	const config = mock<IdentitySubstrateConfig>({
 		trustedKeys,
 		keyRefreshIntervalSeconds: 300,
 		inboundSubjectClaim,
