+++ packages/cli/src/modules/identity-substrate/services/jti-cleanup.service.ts
@@ -4,9 +4,8 @@ import { OnLeaderStepdown, OnLeaderTakeover, OnShutdown } from '@n8n/decorators'
 import { Service } from '@n8n/di';
 import { InstanceSettings } from 'n8n-core';
 
-import { TokenExchangeConfig } from '@/modules/token-exchange/token-exchange.config';
-
 import { TokenExchangeJtiRepository } from '../database/repositories/token-exchange-jti.repository';
+import { IdentitySubstrateConfig } from '../identity-substrate.config';
 
 @Service()
 export class JtiCleanupService {
@@ -18,7 +17,7 @@ export class JtiCleanupService {
 
 	constructor(
 		logger: Logger,
-		private readonly config: TokenExchangeConfig,
+		private readonly config: IdentitySubstrateConfig,
 		private readonly jtiRepository: TokenExchangeJtiRepository,
 		private readonly instanceSettings: InstanceSettings,
 	) {
