+++ packages/cli/src/modules/identity-substrate/services/trusted-key-sync.service.ts
@@ -10,7 +10,10 @@ import { UnexpectedError, UserError, jsonParse } from 'n8n-workflow';
 import { createHash, createPublicKey } from 'node:crypto';
 import { z } from 'zod';
 
-import { TokenExchangeConfig } from '@/modules/token-exchange/token-exchange.config';
+import { TrustedKeySourceEntity } from '../database/entities/trusted-key-source.entity';
+import { TrustedKeyEntity } from '../database/entities/trusted-key.entity';
+import { TrustedKeySourceRepository } from '../database/repositories/trusted-key-source.repository';
+import { IdentitySubstrateConfig } from '../identity-substrate.config';
 import type {
 	JwksKeySource,
 	JwtAlgorithm,
@@ -18,16 +21,12 @@ import type {
 	TrustedKeyData,
 	TrustedKeySource,
 	TrustedKeySourcePolicy,
-} from '@/modules/token-exchange/token-exchange.schemas';
+} from '../identity-substrate.schemas';
 import {
 	TrustedKeySourcePolicySchema,
 	TrustedKeySourceSchema,
-} from '@/modules/token-exchange/token-exchange.schemas';
-
+} from '../identity-substrate.schemas';
 import { JwksResolverService } from './jwks-resolver';
-import { TrustedKeySourceEntity } from '../database/entities/trusted-key-source.entity';
-import { TrustedKeyEntity } from '../database/entities/trusted-key.entity';
-import { TrustedKeySourceRepository } from '../database/repositories/trusted-key-source.repository';
 
 type AlgorithmFamily = 'RSA' | 'EC' | 'EdDSA';
 
@@ -74,7 +73,7 @@ export class TrustedKeySyncService {
 
 	constructor(
 		logger: Logger,
-		private readonly config: TokenExchangeConfig,
+		private readonly config: IdentitySubstrateConfig,
 		private readonly trustedKeySourceRepository: TrustedKeySourceRepository,
 		private readonly instanceSettings: InstanceSettings,
 		private readonly dbLockService: DbLockService,
