+++ packages/cli/test/integration/identity-substrate/trusted-key.repository.integration.test.ts
@@ -7,7 +7,7 @@ import type {
 	TrustedKeyData,
 	TrustedKeySourceStatus,
 	TrustedKeySourceType,
-} from '@/modules/token-exchange/token-exchange.schemas';
+} from '@/modules/identity-substrate/identity-substrate.schemas';
 
 const makeSource = (
 	overrides: Partial<{
